+++ pyseto/versions/v4.py
@@ -11,10 +11,11 @@ from cryptography.hazmat.primitives.asymmetric.ed25519 import (
 
 from ..exceptions import DecryptError, EncryptError, SignError, VerifyError
 from ..key_interface import KeyInterface
+from ..local_key import LocalKey
 from ..utils import base64url_encode, pae
 
 
-class V4Local(KeyInterface):
+class V4Local(LocalKey):
     """
     The key object for v4.local.
     """